ported nojillium to forge 1.18.2
