+++ api/src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+	"pack": {
+		"description": "Nojillium Resources",
+		"pack_format": 8
+	}
+}
