+++ build.gradle.kts
@@ -137,7 +137,7 @@ fun arrayOfProjects(vararg projectNames: String): Array<String> {
     return listOf(*projectNames).filter { p -> findProject(p) != null }.toTypedArray()
 }
 val uniminedImpls = arrayOfProjects(
-    "neoforge",
+    "forge",
 )
 
 allprojects {
@@ -164,10 +164,10 @@ allprojects {
     tasks.withType<JavaCompile> {
         if (name !in arrayOf("compileMcLauncherJava", "compilePatchedMcJava")) {
             options.encoding = "UTF-8"
-            sourceCompatibility = "21"
-            options.release = 21
+            sourceCompatibility = "17"
+            options.release = 17
             javaCompiler = javaToolchains.compilerFor {
-                languageVersion = JavaLanguageVersion.of(21)
+                languageVersion = JavaLanguageVersion.of(17)
             }
             options.compilerArgs.addAll(arrayOf("-Xplugin:Manifold no-bootstrap"))
         }
@@ -195,7 +195,7 @@ allprojects {
             .mapValues { entry -> entry.value as String })
         props["mod_version"] = ZumeGradle.version
 
-        filesMatching(immutableListOf("fabric.mod.json", "META-INF/neoforge.mods.toml")) {
+        filesMatching(immutableListOf("fabric.mod.json", "META-INF/mods.toml")) {
             expand(props)
         }
     }
@@ -217,7 +217,7 @@ subprojects {
     }
 
     dependencies {
-        implementation("dev.nolij:zson:${"zson_version"()}")
+        implementation("dev.nolij:zson:${"zson_version"()}:downgraded-17")
     }
 
     if (implName in uniminedImpls) {
@@ -270,8 +270,8 @@ unimined.minecraft {
 
     runs.off = true
 
-    neoForge {
-        loader("neoforge_version"())
+    minecraftForge {
+        loader("forge_version"())
     }
 
     mappings {
@@ -288,7 +288,7 @@ val shade: Configuration by configurations.creating {
 }
 
 dependencies {
-    shade("dev.nolij:zson:${"zson_version"()}")
+    shade("dev.nolij:zson:${"zson_version"()}:downgraded-17")
 
     compileOnly("org.apache.logging.log4j:log4j-core:${"log4j_version"()}")
 
@@ -303,6 +303,9 @@ dependencies {
 
 tasks.jar {
     enabled = false
+	manifest {
+		attributes["MixinConfigs"] = "nolijium-forge.mixins.json" 
+	}
 }
 
 val sourcesJar = tasks.register<Jar>("sourcesJar") {
@@ -417,7 +420,7 @@ afterEvaluate {
         version = ZumeGradle.version
         changelog = getChangelog()
 
-        modLoaders.addAll("neoforge")
+        modLoaders.addAll("forge")
         dryRun = !isRelease
 
         github {
@@ -432,7 +435,7 @@ afterEvaluate {
                 accessToken = providers.environmentVariable("MODRINTH_TOKEN")
                 projectId = "KstN3eSL"
 
-                minecraftVersions.add("1.21")
+                minecraftVersions.add("1.18.2")
             }
 
             curseforge {
@@ -441,7 +444,7 @@ afterEvaluate {
                 projectId = "969602"
                 projectSlug = "nolijium"
 
-	            minecraftVersions.add("1.21")
+	            minecraftVersions.add("1.18.2")
             }
 
             discord {
