+++ forge/src/main/resources/META-INF/mods.toml
@@ -1,0 +1,36 @@
+modLoader="javafml"
+loaderVersion="[1,)"
+license="OSL-3.0"
+issueTrackerURL="${issue_url}"
+
+[[mods]]
+modId="nolijium"
+version="${mod_version}"
+displayName="${mod_name}"
+displayURL="${mod_url}"
+logoFile="pack.png"
+authors="${nolij}"
+description="""
+${mod_description}
+"""
+
+[[dependencies.nolijium]]
+modId="forge"
+mandatory=true
+versionRange="${forge_neoforge_range}"
+ordering="NONE"
+side="CLIENT"
+
+[[dependencies.nolijium]]
+modId="minecraft"
+mandatory=true
+versionRange="${forge_minecraft_range}"
+ordering="NONE"
+side="CLIENT"
+
+[[dependencies.nolijium]]
+modId="embeddium"
+mandatory=false
+versionRange="${embeddium_range}"
+ordering="AFTER"
+side="CLIENT"
