+++ gradle.properties
@@ -17,22 +17,22 @@ repo_url = https://github.com/Nolij/Nolijium
 issue_url = https://github.com/Nolij/Nolijium/issues
 
 # Platform-Shared
-minecraft_version = 1.21
-parchment_version = 2024.06.23
+minecraft_version = 1.18.2
+parchment_version = 2022.11.06
 
-# NeoForge
-# https://projects.neoforged.net/neoforged/neoforge
-neoforge_minecraft_range = [1.21,)
-neoforge_version = 76-beta
-neoforge_neoforge_range = [21.0.42-beta,)
+# Forge
+# https://files.minecraftforge.net/net/minecraftforge/forge/index_1.18.2.html
+forge_minecraft_range = [1.18.2,)
+forge_version = 40.2.21
+forge_neoforge_range = [40.2.21,)
 
 # Dependencies
 # https://github.com/Nolij/Zume/releases/latest
 zume_version = 1.0.0
 # https://maven.blamejared.com/org/embeddedt/
 ## https://maven.blamejared.com/org/embeddedt/embeddium-1.21/
-embeddium_neoforge_version = 1.0.5-beta.314+1.21
-embeddium_range = [1.0.5,)
+embeddium_neoforge_version = 0.3.18-git.7e86952+mc1.18.2
+embeddium_range = [0.3.18,)
 # https://github.com/Nolij/ZSON/releases/latest
 zson_version = 0.4.1
 # https://mvnrepository.com/artifact/org.apache.logging.log4j/log4j-core
