--- neoforge/src/main/java/dev/nolij/nolijium/integration/embeddium/NolijiumOptionsStorage.java
@@ -1,21 +1,0 @@
-package dev.nolij.nolijium.integration.embeddium;
-
-import dev.nolij.nolijium.impl.Nolijium;
-import dev.nolij.nolijium.impl.config.NolijiumConfigImpl;
-import org.embeddedt.embeddium.api.options.structure.OptionStorage;
-
-public class NolijiumOptionsStorage implements OptionStorage<NolijiumConfigImpl> {
-	
-	private final NolijiumConfigImpl storage = Nolijium.config.clone();
-	
-	@Override
-	public NolijiumConfigImpl getData() {
-		return storage;
-	}
-	
-	@Override
-	public void save() {
-		NolijiumConfigImpl.replace(storage);
-	}
-	
-}
