--- neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -1,39 +1,0 @@
-modLoader="javafml"
-loaderVersion="[1,)"
-license="OSL-3.0"
-issueTrackerURL="${issue_url}"
-
-[[mixins]]
-config="nolijium-neoforge.mixins.json"
-
-[[mods]]
-modId="nolijium"
-version="${mod_version}"
-displayName="${mod_name}"
-displayURL="${mod_url}"
-logoFile="icon.png"
-authors="${nolij}"
-description="""
-${mod_description}
-"""
-
-[[dependencies.nolijium]]
-modId="neoforge"
-type="required"
-versionRange="${neoforge_neoforge_range}"
-ordering="NONE"
-side="CLIENT"
-
-[[dependencies.nolijium]]
-modId="minecraft"
-type="required"
-versionRange="${neoforge_minecraft_range}"
-ordering="NONE"
-side="CLIENT"
-
-[[dependencies.nolijium]]
-modId="embeddium"
-type="optional"
-versionRange="${embeddium_range}"
-ordering="AFTER"
-side="CLIENT"
