--- neoforge/src/main/resources/nolijium-neoforge.mixins.json
@@ -1,17 +1,0 @@
-{
-	"required": true,
-	"minVersion": "0.8",
-	"package": "dev.nolij.nolijium.mixin",
-	"compatibilityLevel": "JAVA_21",
-	"client": [
-		"neoforge.ChatComponentMixin",
-		"neoforge.GameRendererMixin",
-		"neoforge.LevelLightEngineMixin",
-		"neoforge.LevelRendererMixin",
-		"neoforge.LightTextureMixin",
-		"neoforge.TextureAtlasMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
